add models.py to models foldder and rename it redflag.py
